backend/requirements.txt
flask
flask-cors
openai
faiss-cpu
numpy
scikit-learn
python-dotenv
requests
